func: add run_wasm function
The function can be used to run abitrary WebAssembly
functions. It takes the WebAssembly module body
and the function name in user data, and forwards
all the arguments to this function as is.
The final ABI for these functions still remains to be specified,
but for now the following assumptions hold true:
 - integers are passed as is
 - doubles are passed as is
 - strings are passed as a pointer to the following structure:
     [1 byte with value SQLITE_TEXT (3)][data]
 - blobs are going to be passed as a pointer to:
     [1 byte with value SQLITE_BLOB (4)][4 bytes of size][blob]
   ... but are not supported yet
 - nulls are passed as a pointer to:
     [1 byte with value SQLITE_NULL (5)]
   ... but are not supported yet