fiddle: add a small Python server for COOP/COEP
In order to experiment with Origin-Private File System (OPFS),
certain HTTP headers need to be sent from the server to opt-in
into newest features. It's also advised to run bleeding-edge
Chrome, e.g. version 108 seems to support all of it.
To make development easier, a simple server which adds the required
headers is added in fiddle/ directory.