libsql-client: unify with rusqlite type definitions
Our users are likely going to expect the types to match what
rusqlite (and sqlx) already have for SQLite-compatible databases,
so let's unify the type enum.
Since sqld type definitions also match rusqlite, we get that part
of compatibility for free - and that's important, since eventually
we should just use a common crate for client and server types.