metrics: add LIBSQL_STMTSTATUS_ROWS_READ and _WRITTEN
The stats are gathered in order to estimate how many rows
were accessed when executing a statement.
While similar metrics already exist - like fullscan steps
and VM steps - they are not always 1:1 with the number of rows
touched in the process. The built-in count() aggregate
is one of the exceptions. It technically does not read rows,
but instead iterates over B-Tree pages, but it's asymptotically
the same.
For writes, sqlite3_changes() function already exists and informs
how many rows were affected by the most recent insert/update/delete,
but it does not count triggers, replaces, etc into account.