reserve table names (#381)
some tables come from libSQL and cannot be deleted. Others, are usually
associated with backups and internal functions and should also not be
deleted.

A long term better strategy is to allow users to specify which tables
cannot be dropped, but litestream is widespread enough, and simple
enough, that we can assume that if it is in use, we don't want those
tables to be dropped. So a static list works for now.

Co-authored-by: Glauber Costa <glauber@turso.tech>