Fix corruption on replicating transactions with duplicated pages (#494)
* replica: inject frames with writable_schema=on

... and sync the schema changes right after the injection
with writable_schema=reset, as per pragma documentation.

Full tribute and kudos to @MarinPostma for coming up with this one.

* replica: add a PGHDR_DIRTY flag to page headers

It's not really needed right now, but it conforms to the internals
of SQLite, so we should set it just in case the code base starts
relying on it, or asserting its value.

* replica: ensure page headers store pages in frame order

This commit makes sure that page headers are added in the dirty list in the
order of arrival. A database corruption was observed after a sufficiently large
transaction actually had two versions of a single page in a batch. Frames are
applied in dirty-list order, so due to the fact that we previously appended new
frames to the beginning on the list, we ended up writing the newest version of
the page first, and then rewriting it with the old one, hence the
inconsistency. Whew.