HTTP replication (#537)
* replication: add HTTP endpoint skeleton

* replication: add HTTP implementation for /frames endpoint

It still does not perform any kind of handshake, just asks
for frames starting N.

Tested with:
 $ cargo run -- --http-replication-listen-addr 127.0.0.1:8081
 $ curl -d '{"next_offset": 0}' -v localhost:8081/frames

* replication: add /hello endpoint

Serves the same purpose as gRPC hello and provides:
 - generation id
 - generation start index
 - database id

* replication: limit HTTP response in frames to 256

An arbitrary limit to make sure we do not overload memory.

* replication/primary: unconditionally update max frame number

With gRPC replication, it was reasonable to assume that there
are listeners to the max frame number notifier, but with HTTP
it's not necessarily the case. Since watch::send() fails if there are
no receivers, we hereby switch to send_replace(), which successfully
updates the value even if there are no active receivers.

* replication: add crude snapshot support for HTTP

The HTTP replication will now react to SnapshotRequired error
by just sending all frames from a snapshot to the user.
That's prone to overcommitting memory, but better than giving up.
This change should be followed up by streaming the snapshot frames
in multiple smaller bits.

* replication: migrate HTTP to Axum

Following the example in admin_api.rs