Fix the number of semaphore units taken (#587)
* factory: fix the number of semaphore units taken

In manual tests, the value was mistakenly updated to 64,
in order to easily trigger an error - unfortunately, it also
somehow made it to the main branch...

* lib: restore previous concurrency limit of 128

It was reduced to 32 for testing purposes only, but then it unfortunately
sneaked through to the main branch.

* factory: add tracing available semaphore units

Will be useful when debugging overload protection in the future.