write proxy: expire old write delegation streams (#723)
      * write proxy: expire old write delegation streams

Right now we don't have a mechanism for expiring write delegation
streams that disconnected, but without sending a valid Disconnect
messages. That unfortunately also means that we leak libSQL connections,
which are kept in-memory indefinitely. That leaks memory as well as disk
space, because connections also keep temporary tables in place.

FIXME: we also need to keep information about expired streams for longer,
in case a client tries to send a late message. Such a message should be
rejected in order to avoid running it outside of transaction context.

* write proxy: apply review fixes to garbage collection